Endless save bug fixed.
Problem:
IF we tried to get  an enum which is not available any more in enum definition (for example enum entry got renamed), code tried to put default value. put method executes get to validate changes.  -> Endless


Solution:
If we have an exception in enum restoral, we remove the entry directly from the map. 

git-svn-id: svn://svn.appwork.org/utils@977 21714237-3853-44ef-a1f0-ef8f03a7d1fe